requirement.txt
asgiref==3.6.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
Deprecated==1.2.13
Django==4.2.1
django-oauth-toolkit==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
jwcrypto==1.4.2
oauthlib==3.2.2
pycparser==2.21
PyJWT==2.7.0
pytz==2023.3
requests==2.30.0
rest-framework-simplejwt==0.0.2
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.2
wrapt==1.15.0
